Add docker-compose setup for development
To make it easier to setup the whole development environment and avoid
to run a postgres / redis instance on your local development machine.

The docker-compose file will start three containers:
* postgres database (db)
* redis (cache)
* frontend (sidekiq, puma & webpacker)

We use the official postgres, redis and ruby docker images from DockerHub.

When starting the first time the containers, it is necessary to build the
frontend container with docker-compose build. After that, the containers
can get started with docker-compose up.

See updated README file for a more detailed explanation how to use
the docker setup.